feat: upgrade to blockscout v10.1.1 with igra patches
Rebases the igra branch onto upstream v10.1.1 (from v9.0.2).

Igra-specific patches carried forward:
- withdrawal.ex: :integer → :decimal for index and validator_index
  (EIP-4895 uint64 overflow fix, upstream still unfixed)
- Migration to ALTER withdrawal columns to numeric(20,0)
- CI config: add igra branch to pull_request trigger

+++ apps/explorer/lib/explorer/chain/withdrawal.ex
@@ -13,8 +13,8 @@ defmodule Explorer.Chain.Withdrawal do
 
   @primary_key false
   typed_schema "withdrawals" do
-    field(:index, :integer, primary_key: true, null: false)
-    field(:validator_index, :integer, null: false)
+    field(:index, :decimal, primary_key: true, null: false)
+    field(:validator_index, :decimal, null: false)
     field(:amount, Wei, null: false)
 
     belongs_to(:address, Address,
